fastapi-app/requirements.txt
fastapi==0.120.2
uvicorn[standard]==0.38.0
httpx==0.28.1
PyYAML==6.0.2
